backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.21
# psycopg2-binary removed for local Windows installs without PostgreSQL dev libs
# If you need PostgreSQL support, install `psycopg2-binary` separately or install PostgreSQL dev tools.
# psycopg2-binary==2.9.7
google-auth==2.23.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.102.0
python-docx==0.8.11
PyPDF2==3.0.1
openai==0.27.8
# langchain removed for Windows dev to avoid heavy native deps (numpy/build tools).
# Add langchain later in a Linux/conda environment or when wheels are available.
# langchain==0.0.327
# faiss-cpu removed because prebuilt wheels are not always available on Windows.
# Add faiss-cpu later on appropriate Linux environment or use a hosted vector DB.
# faiss-cpu==1.7.4
# pydantic removed for Windows dev (requires Rust/pydantic-core build). Add back in Linux/production or install Rust toolchain.
# pydantic==2.4.2
